requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
slowapi==0.1.9
pydantic==2.11.2
aiosqlite
